chore(dashboard): rename sidebar.user-info.tsx to sidebar-user-info.tsx
